feat: bring features from node-coveralls and Github action (#9)
* ENVs to handle custom parameters
* base-path for LCOV format
* fixes Invalid Int32 #5 - replace Int32 to Int64
* add branches support for simplecov format

+++ src/coverage_reporter/file_report.cr
@@ -5,12 +5,12 @@ module CoverageReporter
 
     def initialize(
       @name : String,
-      @coverage : Array(Int32?),
-      @branches : Array(Int32?) | Nil = nil
+      @coverage : Array(Int64?),
+      @branches : Array(Int64?) | Nil = nil
     )
     end
 
-    def to_h : Hash(Symbol, String | Array(Int32?))
+    def to_h : Hash(Symbol, String | Array(Int64?))
       {
         :name     => @name,
         :coverage => @coverage,
